cassandra_database/requirements.txt
cassandra-driver
python-dotenv
pyyaml
